Removed queue from storing to local data file

+++ lib/services/file_system_service.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'dart:async';
+import 'package:my_pat/service_locator.dart';
 import 'package:my_pat/services/services.dart';
 import 'package:my_pat/utils/log/log.dart';
 import 'package:path_provider/path_provider.dart';
@@ -42,49 +43,56 @@ class FileSystemService {
 
   Future<Response> allocateSpace() async {
     File localFile = await localDataFile;
-    if (await localFile.exists()) {
-      Log.info(TAG,"data file from previous session is found, deleting...");
-      try {
-        await localFile.delete();
-      } catch (e) {
-        Log.shout(TAG,'data file delete failed $e');
-      }
-    }
-    Log.info(TAG,'checking sufficient local storage space...,');
+    final int ln = await localFile.length();
+    print("LOCAL DATA FILO $ln");
+    TestStates testState = sl<SystemStateManager>().testState;
 
-    try {
-      var spaceToAllocate = GlobalSettings.minStorageSpaceMB * 1024;
-      Log.info(TAG,'Free storage space: $spaceToAllocate required');
-      RandomAccessFile file = await localFile.open(mode: FileMode.write);
-      file.truncateSync(spaceToAllocate);
-      file.truncateSync(0);
-      file.closeSync();
+    // todo implement constant space allocation
+    if (testState == TestStates.NOT_STARTED) {
+      if (await localFile.exists()) {
+        Log.info(TAG, "data file from previous session is found, deleting...");
+        try {
+          await localFile.delete();
+        } catch (e) {
+          Log.shout(TAG, 'data file delete failed $e');
+        }
+      }
+      Log.info(TAG, 'checking sufficient local storage space...,');
+
+      try {
+        var spaceToAllocate = GlobalSettings.minStorageSpaceMB * 1024;
+        Log.info(TAG, 'Free storage space: $spaceToAllocate required');
+        RandomAccessFile file = await localFile.open(mode: FileMode.write);
+        file.truncateSync(spaceToAllocate);
+        file.closeSync();
+        return Response(success: true);
+      } catch (e) {
+        Log.shout(TAG, 'SPACE ALLOCATION FAILED $e');
+        return Response(success: false, error: e.toString());
+      }
+    } else {
       return Response(success: true);
-    } catch (e) {
-      Log.shout(TAG,'SPACE ALLOCATION FAILED $e');
-      return Response(success: false, error: e.toString());
     }
   }
 
   Future<Response> init() async {
-
     try {
-      Log.info(TAG,'Attempt to create initial files...');
+      Log.info(TAG, 'Attempt to create initial files...');
       File mainLogFile = await logMainFile;
       await mainLogFile.create();
-      Log.info(TAG,'MAIN_LOG_FILE CREATED');
+      Log.info(TAG, 'MAIN_LOG_FILE CREATED');
       File localFile = await localDataFile;
       await localFile.create();
-      Log.info(TAG,'LOCAL_DATA_FILE CREATED');
+      Log.info(TAG, 'LOCAL_DATA_FILE CREATED');
       File logInFile = await logInputFile;
       logInFile.create();
-      Log.info(TAG,'LOG_INBOUND_FILE CREATED');
+      Log.info(TAG, 'LOG_INBOUND_FILE CREATED');
       File logOutFile = await logInputFile;
       logOutFile.create();
-      Log.info(TAG,'LOG_OUTBOUND_FILE CREATED');
+      Log.info(TAG, 'LOG_OUTBOUND_FILE CREATED');
       return Response(success: true);
     } catch (e) {
-      Log.warning(TAG,'FILES CREATION ERROR: ${e.toString()}');
+      Log.warning(TAG, 'FILES CREATION ERROR: ${e.toString()}');
       return Response(success: false, error: e.toString());
     }
   }
@@ -97,12 +105,11 @@ class FileSystemService {
       logInFile.delete();
       File logOutFile = await logInputFile;
       logOutFile.delete();
-      Log.info(TAG,'FILES CLEARED');
+      Log.info(TAG, 'FILES CLEARED');
       return Response(success: true);
     } catch (e) {
-      Log.warning(TAG,'FILES CLEAR ERROR: ${e.toString()}');
+      Log.warning(TAG, 'FILES CLEAR ERROR: ${e.toString()}');
       return Response(success: false, error: e.toString());
     }
   }
 }
-
